edition hook added
